6.10

+++ src/main/java/com/ovft/configure/sys/service/impl/EduOfflineNumServiceImpl.java
@@ -1,9 +1,6 @@
 package com.ovft.configure.sys.service.impl;
 
-import com.ovft.configure.sys.bean.EduOfflineNum;
-import com.ovft.configure.sys.bean.EduOfflineNumExample;
-import com.ovft.configure.sys.bean.EduOfflineOrderExample;
-import com.ovft.configure.sys.bean.EduOfflineOrderitemExample;
+import com.ovft.configure.sys.bean.*;
 import com.ovft.configure.sys.dao.EduOfflineNumMapper;
 import com.ovft.configure.sys.dao.EduOfflineOrderMapper;
 import com.ovft.configure.sys.dao.EduOfflineOrderitemMapper;
@@ -47,12 +44,24 @@ public class EduOfflineNumServiceImpl implements EduOfflineNumService {
     @Override
     @Transactional
     public Integer deleteOffLineNumCourse(EduOfflineNum eduOfflineNum) {
-        //删除课程记录
-        int res = eduOfflineNumMapper.deleteByPrimaryKey(eduOfflineNum.getId());
+
         //删除订单的记录
         EduOfflineOrderExample eduOfflineOrderExample = new EduOfflineOrderExample();
         eduOfflineOrderExample.createCriteria().andUserIdEqualTo(eduOfflineNum.getUserId()).andCourseIdEqualTo(eduOfflineNum.getCourseId());
         eduOfflineOrderMapper.deleteByExample(eduOfflineOrderExample);
+
+        //删除课程记录
+        //判断是否有这个线下支付订单 如果没有最后删除用户信息
+        //查询出是否还有订单
+        EduOfflineOrderExample eduOfflineOrderExample1 = new EduOfflineOrderExample();
+        eduOfflineOrderExample1.createCriteria().andUserIdEqualTo(eduOfflineNum.getUserId()).andCourseIdEqualTo(eduOfflineNum.getCourseId());
+        List<EduOfflineOrder> eduOfflineOrders = eduOfflineOrderMapper.selectByExample(eduOfflineOrderExample1);
+        int res = 1;
+
+        if (eduOfflineOrders.size() == 1) {
+            res = eduOfflineNumMapper.deleteByPrimaryKey(eduOfflineNum.getId());
+        }
+
         //删除订单详细的记录
         EduOfflineOrderitemExample eduOfflineOrderitemExample = new EduOfflineOrderitemExample();
         eduOfflineOrderitemExample.createCriteria().andUserIdEqualTo(eduOfflineNum.getUserId()).andCourseIdEqualTo(eduOfflineNum.getCourseId());
